+++ src/wetterdienst/provider/meteofrance/synop/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (C) 2018-2025, earthobservations developers.
+# Distributed under the MIT License. See LICENSE for more info.
+from wetterdienst.provider.meteofrance.synop.api import (
+    MeteoFranceSynopMetadata,
+    MeteoFranceSynopRequest,
+)