+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Types of changes:
 
 ## [Unreleased]
 
+- Add Météo-France (France) synop network (subdaily, 3-hourly)
 - Add MeteoSwiss (Switzerland) observation provider with 10-minute, hourly, daily, monthly and annual resolution
 
 ## [0.127.0] - 2026-07-07